refactor: split terminal session logic into separate files
